fix(mold,cook): harden handoff and bundle boundaries
