+++ .hallouminate/wiki/gotchas/codeql-generated-file-implicit-concat.md
@@ -1,3 +1,3 @@
 # CodeQL implicit-string-concat on pprint-generated files
 
-`pprint.pformat` wraps long strings into adjacent implicit-concat fragments whenever they exceed `width` at their nesting depth — CodeQL's `py/implicit-string-concatenation-in-list` flags every fragment pair. For `src/mold/_document_rules.py` the first clean width was measured at **161** (80→45 pairs, 150→1, 160→1, 161→0); the clean width is a function of the longest projected string at its indent, so a longer `CrossFieldRule.description` silently reintroduces alerts. Guard: `test_document_rules_compiler.py` asserts no generated line ends in a quote with the next beginning in one. For hand-written wrapped strings in list literals, parenthesize the group — the upstream query excludes `isParenthesized()` matches. `ruff.toml` selects only E4/E7/E9/F, so long generated lines are lint-safe.
+`pprint.pformat` wraps long strings into adjacent implicit-concat fragments whenever they exceed `width` at their nesting depth — CodeQL's `py/implicit-string-concatenation-in-list` flags every fragment pair. For `src/easy_cheese/skills/mold/_document_rules.py` the first clean width was measured at **161** (80→45 pairs, 150→1, 160→1, 161→0); the clean width is a function of the longest projected string at its indent, so a longer `CrossFieldRule.description` silently reintroduces alerts. Guard: `test_document_rules_compiler.py` asserts no generated line ends in a quote with the next beginning in one. For hand-written wrapped strings in list literals, parenthesize the group — the upstream query excludes `isParenthesized()` matches. `ruff.toml` selects only E4/E7/E9/F, so long generated lines are lint-safe.
