--- src/easy_cheese/shared/legacy_handoffs.py
@@ -1,5 +1,0 @@
-"""Exact-version-only legacy handoff migration."""
-
-from easy_cheese.shared.handoffs import LegacyHandoff, migrate
-
-__all__ = ["LegacyHandoff", "migrate"]
