+++ src/easy_cheese/skills/cook/commands.py
@@ -1,13 +1,31 @@
 """Cook's declarative bundle commands and canonical consumer boundary."""
 
+# Transitional shared modules are imported after a source-tree path fallback.
+# ruff: noqa: E402
+
 from __future__ import annotations
 
 import argparse
-import importlib
 import json
 import sys
 from pathlib import Path
 from typing import Any
+from urllib.parse import unquote, urlsplit
+
+_SHARED_SOURCE = Path(__file__).resolve().parents[4] / "shared" / "scripts"
+if _SHARED_SOURCE.is_dir() and str(_SHARED_SOURCE) not in sys.path:
+    sys.path.insert(0, str(_SHARED_SOURCE))
+
+import cli as _cli
+import findings_cli as _findings_cli
+import gates_cli as _gates_cli
+import handoff_cli as _handoff_cli
+import html_report_cli as _html_report_cli
+import paths_cli as _paths_cli
+import read_handoff_slug as _read_handoff_slug
+import slugify as _slugify
+import worktree as _worktree
+import write_handoff_artifact as _write_handoff_artifact
 
 from easy_cheese.shared.artifact_path import main as _artifact_path_main
 from easy_cheese.shared.bundle_commands import bundle_command, dispatch
@@ -18,6 +36,7 @@ from easy_cheese.shared.handoffs import (
     accept,
     canonical_bytes,
     pointer_from_mapping,
+    read_nofollow_file,
 )
 from easy_cheese_schemas.schema_runtime import (
     SCHEMA_ROOT,
@@ -29,31 +48,43 @@ from easy_cheese_schemas.schema_runtime import (
 )
 
 
-def contract_accept(pointer: HandoffPointer) -> AcceptedArtifact:
+def contract_accept(pointer: HandoffPointer, expected_root: Path) -> AcceptedArtifact:
     """Accept a route-bound canonical Mold-to-Cook handoff."""
-    return accept(pointer)
+    return accept(pointer, expected_root)
 
 
-def execute(pointer: HandoffPointer) -> Any:
+def execute(pointer: HandoffPointer, expected_root: Path) -> Any:
     """Return only the canonical plan produced by full pointer acceptance."""
-    return contract_accept(pointer).canonical
+    return contract_accept(pointer, expected_root).canonical
 
 
 def _read_pointer(path: Path) -> HandoffPointer:
-    raw = json.loads(path.read_text(encoding="utf-8"))
+    if path.is_symlink():
+        raise HandoffError("pointer file must not be a symlink")
+    data = read_nofollow_file(path.parent, path.name)
+    raw = json.loads(data)
     if not isinstance(raw, dict):
         raise HandoffError("pointer must be an object")
     pointer = pointer_from_mapping(raw)
-    expected = (path.parent.parent / "pointers" / f"{pointer.operation_id}.json").resolve()
+    if canonical_bytes(pointer) != data:
+        raise HandoffError("pointer must use canonical JSON")
+    if path.parent.is_symlink():
+        raise HandoffError("pointer operation directory must not be a symlink")
+    operation_root = path.parent.parent.resolve()
+    payload = urlsplit(pointer.payload.uri)
+    if payload.scheme != "file" or payload.netloc not in {"", "localhost"}:
+        raise HandoffError("pointer payload must use a local file URI")
+    expected_payload = operation_root / "payloads" / f"{pointer.operation_id}.json"
+    if Path(unquote(payload.path)).resolve() != expected_payload:
+        raise HandoffError("pointer payload escapes the operation root")
+    expected = (operation_root / "pointers" / f"{pointer.operation_id}.json").resolve()
     if path.resolve() != expected:
         raise HandoffError("pointer path does not match its operation id")
     return pointer
 
 
-def _shared_command(module_name: str, argv: list[str]) -> int:
-    cli = importlib.import_module("cli")
-    module = importlib.import_module(module_name)
-    return cli.run(module._setup, argv=argv)
+def _shared_command(module: Any, argv: list[str]) -> int:
+    return _cli.run(module._setup, argv=argv)
 
 
 @bundle_command("artifact-path")
@@ -65,57 +96,55 @@ def artifact_path_command(argv: list[str]) -> int:
 @bundle_command("worktree")
 def worktree_command(argv: list[str]) -> int:
     """Create, harvest, or tear down a curd worktree."""
-    cli = importlib.import_module("cli")
-    module = importlib.import_module("worktree")
-    return cli.run(module._setup, argv=argv)
+    return _cli.run(_worktree._setup, argv=argv)
 
 
 @bundle_command("slugify")
 def slugify_command(argv: list[str]) -> int:
     """Derive a bounded task slug and artifact path."""
-    return _shared_command("slugify", argv)
+    return _shared_command(_slugify, argv)
 
 
 @bundle_command("write_handoff_artifact")
 def write_handoff_artifact_command(argv: list[str]) -> int:
     """Write a canonical phase handoff artifact atomically."""
-    return _shared_command("write_handoff_artifact", argv)
+    return _shared_command(_write_handoff_artifact, argv)
 
 
 @bundle_command("read_handoff_slug")
 def read_handoff_slug_command(argv: list[str]) -> int:
     """Read and validate a canonical handoff preamble."""
-    return _shared_command("read_handoff_slug", argv)
+    return _shared_command(_read_handoff_slug, argv)
 
 
 @bundle_command("findings_cli")
 def findings_command(argv: list[str]) -> int:
     """Render or select structured review findings."""
-    return _shared_command("findings_cli", argv)
+    return _shared_command(_findings_cli, argv)
 
 
 @bundle_command("gates_cli")
 def gates_command(argv: list[str]) -> int:
     """Classify readiness from the quality scoreboard."""
-    return _shared_command("gates_cli", argv)
+    return _shared_command(_gates_cli, argv)
 
 
 @bundle_command("paths_cli")
 def paths_command(argv: list[str]) -> int:
     """Resolve and validate workflow artifact paths."""
-    return _shared_command("paths_cli", argv)
+    return _shared_command(_paths_cli, argv)
 
 
 @bundle_command("handoff_cli")
 def handoff_command(argv: list[str]) -> int:
     """Render, parse, or dispatch a canonical handoff preamble."""
-    return _shared_command("handoff_cli", argv)
+    return _shared_command(_handoff_cli, argv)
 
 
 @bundle_command("render_html")
 def render_html_command(argv: list[str]) -> int:
     """Render a Markdown artifact as a self-contained HTML report."""
-    return _shared_command("html_report_cli", argv)
+    return _shared_command(_html_report_cli, argv)
 
 
 @bundle_command("normalize")
@@ -171,7 +200,7 @@ def contract_command(argv: list[str]) -> int:
     accept_parser.add_argument("--pointer", required=True, type=Path)
     arguments = parser.parse_args(argv)
     pointer = _read_pointer(arguments.pointer)
-    accepted = contract_accept(pointer)
+    accepted = contract_accept(pointer, arguments.pointer.parent.parent)
     sys.stdout.buffer.write(canonical_bytes(accepted.canonical))
     sys.stdout.buffer.write(b"\n")
     return 0
