+++ src/easy_cheese/skills/mold/validate_spec.py
